Move everything test-environment related into ./environment
No functional changes were made in this patch.
