+++ api_tests/jest.config.js
@@ -6,10 +6,12 @@ module.exports = {
         "^.+\\.(t|j)s$": "ts-jest",
     },
     moduleFileExtensions: ["ts", "js", "json", "node"],
-    testEnvironment: "<rootDir>/dist/src/test_environment",
+    testEnvironment: "<rootDir>/dist/src/environment/test_environment",
     globalSetup: "<rootDir>/src/environment/setup.ts",
     globalTeardown: "<rootDir>/src/environment/teardown.ts",
-    setupFilesAfterEnv: ["<rootDir>/src/configure_jasmine.ts"],
+    setupFilesAfterEnv: [
+        "<rootDir>/src/environment/jasmine_capture_current_testname.ts",
+    ],
     testTimeout: 123000,
     bail: true,
 };
