+++ api_tests/src/environment/cnd_instance.ts
@@ -2,11 +2,12 @@ import { JsonMap, stringify } from "@iarna/toml";
 import { ChildProcess, spawn } from "child_process";
 import tempWrite from "temp-write";
 import { CndConfigFile } from "../config";
-import { openAsync, sleep } from "../utils";
-import waitForLogMessage from "../wait_for_log_message";
+import { sleep } from "../utils";
+import waitForLogMessage from "./wait_for_log_message";
 import { Logger } from "log4js";
 import path from "path";
-import { crashListener } from "../crash_listener";
+import { crashListener } from "./crash_listener";
+import { openAsync } from "./async_fs";
 
 export class CndInstance {
     private process: ChildProcess;
