+++ api_tests/src/environment/setup.ts
@@ -1,7 +1,7 @@
-import { rimrafAsync } from "../utils";
 import { promises as asyncFs } from "fs";
 import path from "path";
 import killNodes from "./kill_nodes";
+import { rimrafAsync } from "./async_fs";
 
 export default async (config: any) => {
     const root = config.rootDir;
